requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
transformers>=4.30.0
torch>=1.13.0
pydantic>=2.0
pytorch-crf==0.7.2
